third_party/blink/web_tests/platform/generic/fast/events/pointerevents/multi-pointer-preventdefault-expected.txt

Verifies the effect of pointer event prevent-default on pointerdown & mouse/pen events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

--- test with preventDefault on non-primary none ---
- start with mouse, primary pen and non-primary pen outside target -
- move mouse into target & press -
1. target received pointerover primary mouse
2. target received pointerenter primary mouse
3. target received mouseover
4. target received mouseenter
5. target received pointermove primary mouse
6. target received mousemove
7. target received pointerdown primary mouse
8. target received mousedown
- move primary pen into target & press -
9. target received pointerover primary pen
10. target received pointerenter primary pen
11. target received pointermove primary pen
12. target received mousemove
13. target received pointerdown primary pen
14. target received mousedown
- move non-primary pen into target & press -
15. target received pointerover non-primary pen
16. target received pointerenter non-primary pen
17. target received pointermove non-primary pen
18. target received pointerdown non-primary pen
- jiggle mouse in target -
19. target received pointermove primary mouse
20. target received mousemove
- jiggle non-primary pen in target -
21. target received pointermove non-primary pen
- jiggle primary pen in target -
22. target received pointermove primary pen
23. target received mousemove
- release mouse -
24. target received pointerup primary mouse
25. target received mouseup
- release non-primary pen -
26. target received pointerup non-primary pen
- release primary pen -
27. target received pointerup primary pen
28. target received mouseup


--- test with preventDefault on primary mouse ---
- start with mouse, primary pen and non-primary pen outside target -
1. target received pointerout primary mouse
2. target received pointerleave primary mouse
3. target received mouseout
4. target received mouseleave
5. target received pointerout primary pen
6. target received pointerleave primary pen
7. target received pointerout non-primary pen
8. target received pointerleave non-primary pen
- move mouse into target & press -
9. target received pointerover primary mouse
10. target received pointerenter primary mouse
11. target received mouseover
12. target received mouseenter
13. target received pointermove primary mouse
14. target received mousemove
15. target received pointerdown primary mouse, prevented default
- move primary pen into target & press -
16. target received pointerover primary pen
17. target received pointerenter primary pen
18. target received pointermove primary pen
19. target received mousemove
20. target received pointerdown primary pen
21. target received mousedown
- move non-primary pen into target & press -
22. target received pointerover non-primary pen
23. target received pointerenter non-primary pen
24. target received pointermove non-primary pen
25. target received pointerdown non-primary pen
- jiggle mouse in target -
26. target received pointermove primary mouse
- jiggle non-primary pen in target -
27. target received pointermove non-primary pen
- jiggle primary pen in target -
28. target received pointermove primary pen
29. target received mousemove
- release mouse -
30. target received pointerup primary mouse
- release non-primary pen -
31. target received pointerup non-primary pen
- release primary pen -
32. target received pointerup primary pen
33. target received mouseup


--- test with preventDefault on primary pen ---
- start with mouse, primary pen and non-primary pen outside target -
1. target received pointerout primary mouse
2. target received pointerleave primary mouse
3. target received mouseout
4. target received mouseleave
5. target received pointerout primary pen
6. target received pointerleave primary pen
7. target received pointerout non-primary pen
8. target received pointerleave non-primary pen
- move mouse into target & press -
9. target received pointerover primary mouse
10. target received pointerenter primary mouse
11. target received mouseover
12. target received mouseenter
13. target received pointermove primary mouse
14. target received mousemove
15. target received pointerdown primary mouse
16. target received mousedown
- move primary pen into target & press -
17. target received pointerover primary pen
18. target received pointerenter primary pen
19. target received pointermove primary pen
20. target received mousemove
21. target received pointerdown primary pen, prevented default
- move non-primary pen into target & press -
22. target received pointerover non-primary pen
23. target received pointerenter non-primary pen
24. target received pointermove non-primary pen
25. target received pointerdown non-primary pen
- jiggle mouse in target -
26. target received pointermove primary mouse
27. target received mousemove
- jiggle non-primary pen in target -
28. target received pointermove non-primary pen
- jiggle primary pen in target -
29. target received pointermove primary pen
- release mouse -
30. target received pointerup primary mouse
31. target received mouseup
- release non-primary pen -
32. target received pointerup non-primary pen
- release primary pen -
33. target received pointerup primary pen


--- test with preventDefault on non-primary pen ---
- start with mouse, primary pen and non-primary pen outside target -
1. target received pointerout primary mouse
2. target received pointerleave primary mouse
3. target received mouseout
4. target received mouseleave
5. target received pointerout primary pen
6. target received pointerleave primary pen
7. target received pointerout non-primary pen
8. target received pointerleave non-primary pen
- move mouse into target & press -
9. target received pointerover primary mouse
10. target received pointerenter primary mouse
11. target received mouseover
12. target received mouseenter
13. target received pointermove primary mouse
14. target received mousemove
15. target received pointerdown primary mouse
16. target received mousedown
- move primary pen into target & press -
17. target received pointerover primary pen
18. target received pointerenter primary pen
19. target received pointermove primary pen
20. target received mousemove
21. target received pointerdown primary pen
22. target received mousedown
- move non-primary pen into target & press -
23. target received pointerover non-primary pen
24. target received pointerenter non-primary pen
25. target received pointermove non-primary pen
26. target received pointerdown non-primary pen, prevented default
- jiggle mouse in target -
27. target received pointermove primary mouse
28. target received mousemove
- jiggle non-primary pen in target -
29. target received pointermove non-primary pen
- jiggle primary pen in target -
30. target received pointermove primary pen
31. target received mousemove
- release mouse -
32. target received pointerup primary mouse
33. target received mouseup
- release non-primary pen -
34. target received pointerup non-primary pen
- release primary pen -
35. target received pointerup primary pen
36. target received mouseup


PASS successfullyParsed is true

TEST COMPLETE
